scute: Remove prepended nul byte in signature data
* src/agent.c (pksign_parse_result): Check for nul byte prepended
by the agent to the signature value.
--

GPG Agent may prepend a nul byte in the signature value if the
first byte of the signature has its most significant bit set, to
prevent it from being interpreted as a sign bit (see the function
agent_pksign_do, in GnuPG's agent/pksign.c file).

The current sexp parser in Scute does not expect this extra nul
byte, and will reject any signature containing it with a
GPG_ERR_INV_LENGTH error.

This patch checks for an initial nul byte in the signature
data, and removes it.

Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>