Do not request certificate data when LEARNing.
* src/agent.c (scute_agent_learn): Use LEARN --sendinfo instead
of LEARN --send.
--

Using --send is useless here and we don't even collect the data
that the agent may send (no data callback), which may cause the
LEARN command to fail.

Of note, the --sendinfo option is available since GnuPG 2.1.1,
meaning Scute would not work anymore with GnuPG 2.0.x. Since that
branch is now EOLed, this should not be a concern.

Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>