Change license from GPL with Mozilla exception to LGPL.
--
This change allows the use of Scute with software which is not from
Mozilla.

Almost all relevant code was written by employees or contractors of
g10 Code GmbH, who is the original author, and the author of this
patch is the responsible officer hat g10 Code to take such decisions.

The file src/sexp-parse.h comes from GnuPG and was entirely written
and changed by employees of g10 Code.  Thus the change from
LGPv3-or-GPLv2 also to LGPLv2.1 can also be done by the author of this
patch.

The only other major changes (execpt for tweaks of the build system)
came from the current maintainer, Damien Goutte-Gattat, who agreed to
the license change by signed mail dated 2019-02-18 (message-id
20190218165051.dosouwb4d4mlypwn@CHS-TMB-078.qmcr.qmul.ac.uk):

  I hereby agree to have all the code I contributed to the Scute project
  so far re-licensed under the terms of the GNU LGPL, version 2.1 or
  later.

Signed-off-by: Werner Koch <wk@gnupg.org>