build: Do not build doc if tools are missing.
* configure.ac (YAT2M): Check for yat2m helper tool.
(EPSTOPDF): Check removed.
(BUILD_DOC): Enable only if convert and yat2m are present.
--

This patch updates the list of tools to check in order to build
the documentation (add yat2m, remove epstopdf which is no longer
required) and make sure that we do not attempt to build said
documentation if these tools are missing.

A note is also added to the README to help packagers identifying
the build-time dependencies.

GnuPG-bug-id: 5358
Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>