Add no-chain option.
This is the same patch as my previous one, but this one applies
cleanly to the t6002 branch, for convenience.

-- >8 --
Subject: [PATCH scute] Add option to return leaf certificate only.

* src/options.h (_scute_opt_t): Add no_chain.
* src/readconf.c (_scute_read_conf): Set no_chain.
* src/agent.c (scute_gpgsm_get_cert): Ask for chain of certificates
depending on no_chain value.
--

The default behaviour of Scute of returning the full certificate
chain may cause various issues in client applications, some of
which seemingly only expect a single certificate. Thus we make
this behaviour configurable.

Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>