doc: Add Evolution setup information to Scute manual.
--

I've been using Scute with Evolution for a while, and since it uses
Mozilla's NSS library for S/MIME operations anyway it should prove
just as reliable for others. Information on adding PKCS #11 modules to
Evolution has been scarce, so let's give it a mention here.

Fun fact: Evolution supports signing emails with both OpenPGP and
S/MIME simultaneously using two MIME signature parts on the same
mail. This unusual setup could cause the user to be prompted for a PIN
twice, once for OpenPGP and once for S/MIME, so this actually makes
sense.

Signed-off-by: John Scott <jscott@posteo.net>